@@ -12,16 +12,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.1.13",
-    "@angular/common": "~9.1.13",
-    "@angular/compiler": "~9.1.13",
-    "@angular/core": "~9.1.13",
-    "@angular/forms": "~9.1.13",
-    "@angular/localize": "9.1.13",
-    "@angular/platform-browser": "~9.1.13",
-    "@angular/platform-browser-dynamic": "~9.1.13",
-    "@angular/router": "~9.1.13",
-    "@fortawesome/angular-fontawesome": "0.6.1",
+    "@angular/animations": "~10.2.5",
+    "@angular/common": "~10.2.5",
+    "@angular/compiler": "~10.2.5",
+    "@angular/core": "~10.2.5",
+    "@angular/forms": "~10.2.5",
+    "@angular/localize": "10.2.5",
+    "@angular/platform-browser": "~10.2.5",
+    "@angular/platform-browser-dynamic": "~10.2.5",
+    "@angular/router": "~10.2.5",
+    "@fortawesome/angular-fontawesome": "~0.7.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-regular-svg-icons": "5.15.3",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
@@ -30,19 +30,17 @@
     "@pt/pt-topology": "file:./tmp/pt-topology-0.1.0.tgz",
     "bootstrap": "5.0.0-beta2",
     "bootstrap-icons": "1.4.0",
-    "core-js": "^2.5.4",
     "ng-bootstrap-form-validation": "9.0.1",
     "ngx-captcha": "^8.0.1",
     "rxjs": "~6.6.6",
-    "tslib": "^1.10.0",
+    "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.15",
-    "@angular/cdk": "11.2.9",
-    "@angular/cli": "~9.1.15",
-    "@angular/compiler-cli": "~9.1.13",
-    "@angular/language-service": "~9.1.13",
+    "@angular-devkit/build-angular": "~0.1002.4",
+    "@angular/cdk": "~10.2.7",
+    "@angular/cli": "~10.2.4",
+    "@angular/compiler-cli": "~10.2.5",
     "@swimlane/ngx-charts": "17.0.1",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
@@ -50,20 +48,20 @@
     "ag-grid-angular": "22.1.0",
     "ag-grid-community": "22.1.0",
     "codelyzer": "^5.1.2",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "ngx-chips": "2.2.2",
     "ngx-clipboard": "13.0.1",
     "ngx-lightbox": "2.3.0",
-    "protractor": "~5.4.0",
+    "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.8.3",
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.8",
     "vis": "git+https://github.com/PANTHEONtech/vis#develop"
   }
 }
