@@ -17,7 +17,7 @@ import { FormsModule } from '@angular/forms';
     ]
 })
 export class AppAgGridModule {
-    static withComponents(components?: any): ModuleWithProviders {
+    static withComponents(components?: any): ModuleWithProviders<AppAgGridModule> {
         return {
             ngModule: AppAgGridModule,
             providers: [
